clean kode - ca ca

+++ trunk/Node2/motor.h
@@ -12,7 +12,7 @@
 void motor_init(void);
 void motor_reset(void);
 void motor_set_reference(int8_t);
-int8_t motor_get_position(void);
+int motor_get_position(void);
 void motor_set_input(int8_t input);
 void motor_regulator(void);
 ////void motor_DAC_send_loop(void);
